Refactored AddendumContainer component in preparation for co-signer support
