+++ client/src/js/addendum/AddendumContainer.jsx
@@ -1,11 +1,9 @@
 import React, { useState } from 'react'
 import PropTypes from 'prop-types'
 import { makeStyles } from '@material-ui/core/styles'
-import { Addendum, AddendumType } from '../../common/model/addendum'
+import { AddendumType } from '../../common/model/addendum'
 import {
-  Box,
   Button,
-  Card,
   Dialog,
   DialogActions,
   DialogContent,
@@ -13,13 +11,10 @@ import {
   DialogTitle,
   Grid
 } from '@material-ui/core'
-import { Agreement, AgreementType } from '../../common/model/agreement'
-import IdentityForm from '../identity/IdentityForm'
-import IdentityCard from './IdentityCard'
-import * as _ from 'lodash'
+import { Agreement } from '../../common/model/agreement'
 import KeyboardBackspaceIcon from '@material-ui/icons/KeyboardBackspace'
 import { Link, useHistory } from 'react-router-dom'
-import { Alert } from '@material-ui/lab'
+import AddendumForm from './AddendumForm'
 
 const useStyles = makeStyles(theme => ({
   root: {
@@ -29,115 +24,14 @@ const useStyles = makeStyles(theme => ({
 }))
 
 /**
- * This class contains the logic for the AddendumContainer component
- */
-class AddendumContainerCtrl {
-  /**
-   * Check if a User owns this agreement. For now this is true only when the User is the Signer of the Agreement.
-   * This may change when we support co-signers/managers.
-   * @param user {{email: string}} The Firebase user object
-   * @param agreement {Agreement} An Agreement object
-   * @returns {boolean}
-   */
-  static isOwner (user, agreement) {
-    return user.email === agreement.signer.value
-  }
-}
-
-/**
  * Component which given an Agreement displays a list of associated addendums
  */
 function AddendumContainer (props) {
 
   const history = useHistory()
   const classes = useStyles()
-  const [addendums, setAddendums] = useState([])
-  const [activeIdentities, setActiveIdentities] = useState([])
-  const [addedIdentities, setAddedIdentities] = useState([])
-  const [removedIdentities, setRemovedIdentities] = useState([])
-  const [updateInProgress, setUpdateInProgress] = useState(false)
-  const [lastAddendum, setLastAddendum] = useState(null)
+  const [contributorUpdateStatus, setContributorUpdateStatus] = useState(false)
   const [openDialog, setOpenDialog] = useState(false)
-  const [error, setError] = useState(null)
-
-  React.useEffect(() => {
-    props.agreement.getAddendums()
-      .then(setAddendums)
-      .catch(console.error)
-  }, [props.agreement])
-
-  React.useEffect(() => {
-    props.agreement.getWhitelist()
-      .then((res) => {
-        setActiveIdentities(res)
-      })
-  }, [props.agreement.id, addendums])
-
-  React.useEffect(() => {
-    // FIXME this seems to be state more than effect, doublecheck
-    if (addendums && addendums.length) {
-      setLastAddendum(addendums[addendums.length - 1])
-    } else {
-      setLastAddendum(null)
-    }
-  }, [addendums])
-
-  const createAddendum = () => {
-    const addendum = new Addendum(
-      AddendumType.CONTRIBUTOR,
-      props.agreement.id,
-      props.agreement.signer,
-      addedIdentities,
-      removedIdentities
-    )
-
-    setUpdateInProgress(true)
-    addendum.save()
-      .then(res => {
-        setAddedIdentities([])
-        setRemovedIdentities([])
-      })
-      .catch(err => {
-        if (err.code === 'permission-denied') {
-          setError('Permission denied, please try again later')
-          return
-        }
-        setError('Request failed, please try again later')
-      })
-      .finally(() => {
-        setUpdateInProgress(false)
-        setAddendums(addendums => [...addendums, addendum])
-      })
-  }
-
-  const setAddedIdentity = (identity) => {
-    setAddedIdentities(addedIdentities => [identity, ...addedIdentities])
-  }
-
-  const setRemovedIdentity = (identity) => {
-    return (evt) => {
-      evt.preventDefault()
-      setRemovedIdentities(removedIdentities => [identity, ...removedIdentities])
-      _.remove(activeIdentities, identity)
-      setActiveIdentities(activeIdentities)
-    }
-  }
-
-  const undoRemove = (identity) => {
-    return (evt) => {
-      evt.preventDefault()
-      _.remove(removedIdentities, identity)
-      setRemovedIdentities(removedIdentities)
-      setActiveIdentities(activeIdentities => [identity, ...activeIdentities])
-    }
-  }
-
-  const undoAdd = (identity) => {
-    return (evt) => {
-      evt.preventDefault()
-      setAddedIdentities(_.without(addedIdentities, identity))
-    }
-  }
 
   const confirmLeave = (val) => {
     if (val === false) {
@@ -150,8 +44,7 @@ function AddendumContainer (props) {
   const goBack = () => {
     // NOTE we need to use a callback instead of a Link as we want to check
     // that there are no unsaved changes
-
-    if (addedIdentities.length > 0 || removedIdentities.length > 0) {
+    if (contributorUpdateStatus) {
       setOpenDialog(true)
       return
     }
@@ -159,82 +52,16 @@ function AddendumContainer (props) {
     history.goBack()
   }
 
-  const updateForm = (
-    <div className="AddendumContainer__update-form">
-      <Grid item xs={12}>
-        <Grid container spacing={2}>
-
-          <Grid item xs={12}>
-            <h2>Update Agreement</h2>
-            <Box>
-              <p>
-                You can sign a new "addendum" to modify the identities allowed
-                to contribute under this agreement. Use the form below to add
-                identities, or select one from the above list to remove it.
-              </p>
-              <p>
-                Once done, make sure to click on &quot;Sign
-                Addendum&quot; below to apply your changes.
-              </p>
-            </Box>
-          </Grid>
-          {removedIdentities.map((a, i) =>
-            <Grid key={`container-removed-${i}`} item xs={12} sm={12} md={6} lg={4}>
-              <IdentityCard key={i} identity={a} callback={undoRemove} type={'removed'}/>
-            </Grid>
-          )}
-          {addedIdentities.map((a, i) =>
-            <Grid key={`container-added-${i}`} item xs={12} sm={12} md={6} lg={4}>
-              <IdentityCard key={i} identity={a} callback={undoAdd} type={'added'}/>
-            </Grid>
-          )}
-        </Grid>
-        <Card variant='outlined' className={classes.root}>
-          <IdentityForm
-            callback={setAddedIdentity}
-            name={props.agreement.type === AgreementType.INDIVIDUAL ? props.agreement.signer.name : null}/>
-        </Card>
-      </Grid>
-      <Grid item xs={12}>
-        {error ? <Alert severity='error'>{error}</Alert> : null}
-      </Grid>
-      <Grid item xs={12}>
-        <Button
-          fullWidth
-          variant='contained'
-          size='large'
-          color='primary'
-          disabled={updateInProgress || addedIdentities.length + removedIdentities.length === 0}
-          onClick={createAddendum}>
-          Sign Addendum
-        </Button>
-      </Grid>
-    </div>
-  )
   return (
     <Grid container spacing={2} className="AddendumContainer">
-      <Grid item xs={12} className="AddendumContainer__active-identities">
-        <h2>Active Identities for this Agreement</h2>
-        <p>Here is a list of identities that are authorized to contribute code
-          under this agreement: {activeIdentities.length === 0 ? <strong>EMPTY</strong> : ''}</p>
-        <Grid container spacing={2}>
-          {activeIdentities.map((a, i) =>
-            <Grid key={`container-${i}`} item xs={12} sm={12} md={6} lg={4} className="AddendumContainer__active-identity">
-              <IdentityCard key={i} identity={a}
-                            callback={setRemovedIdentity} type={'default'}/>
-            </Grid>
-          )}
-        </Grid>
+      <Grid item xs={12}>
+        <AddendumForm
+          user={props.user}
+          agreement={props.agreement}
+          addendumType={AddendumType.CONTRIBUTOR}
+          updateStatus={setContributorUpdateStatus}
+        />
       </Grid>
-      <Grid item xs={12} className="AddendumContainer__summary">
-        <p>
-          We have {addendums ? addendums.length : 0} addendums on file for this
-          agreement. The last one was signed
-          on: {lastAddendum ? lastAddendum.dateSigned.toString() : 'NEVER'}</p>
-      </Grid>
-      {/* TODO print a list of all the addendums if it's admin */}
-      {AddendumContainerCtrl.isOwner(props.user, props.agreement) ? updateForm : null}
-
       <Grid item xs={12}>
         <Dialog
           onClose={confirmLeave}
